Переименовывает папку с тестовыми данными
