+++ project/src/fixture/favorite-offers.ts
@@ -1,37 +1,8 @@
-import {cities} from 'mocks/cities';
-import {hosts} from 'mocks/hosts';
+import {cities} from 'fixture/cities';
+import {hosts} from 'fixture/hosts';
 import {Offers} from 'types/offer';
 
-const offers: Offers = [
-  {
-    bedrooms: 3,
-    city: cities[0],
-    description: 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    goods: [
-      'Heating',
-    ],
-    host: hosts[0],
-    id: 1,
-    images: [
-      'http://picsum.photos/260/200?r=3',
-      'http://picsum.photos/260/200?r=4',
-      'http://picsum.photos/260/200?r=5',
-      'http://picsum.photos/260/200?r=6',
-    ],
-    isFavorite: true,
-    isPremium: false,
-    location: {
-      latitude: 52.35514938496378,
-      longitude: 4.673877537499948,
-      zoom: 8,
-    },
-    maxAdults: 4,
-    previewImage: 'http://picsum.photos/260/200?r=1',
-    price: 120,
-    rating: 3.5,
-    title: 'Beautiful & luxurious studio at great location',
-    type: 'Apartment',
-  },
+const favoriteOffers: Offers = [
   {
     bedrooms: 2,
     city: cities[0],
@@ -48,7 +19,7 @@ const offers: Offers = [
       'http://picsum.photos/260/200?r=11',
       'http://picsum.photos/260/200?r=12',
     ],
-    isFavorite: false,
+    isFavorite: true,
     isPremium: true,
     location: {
       latitude: 52.55514938496378,
@@ -60,7 +31,7 @@ const offers: Offers = [
     price: 320,
     rating: 5.0,
     title: 'Luxurious studio',
-    type: 'Private room',
+    type: 'Hostel',
   },
   {
     bedrooms: 1,
@@ -76,7 +47,7 @@ const offers: Offers = [
       'http://picsum.photos/260/200?r=16',
     ],
     isFavorite: false,
-    isPremium: true,
+    isPremium: false,
     location: {
       latitude: 52.55514938496378,
       longitude: 4.573877537499948,
@@ -91,7 +62,7 @@ const offers: Offers = [
   },
   {
     bedrooms: 2,
-    city: cities[0],
+    city: cities[1],
     description: 'Flat in a beautiful old town, middle-eastern paradise.',
     goods: [
       'Heating',
@@ -104,7 +75,7 @@ const offers: Offers = [
       'http://picsum.photos/260/200?r=10',
     ],
     isFavorite: true,
-    isPremium: false,
+    isPremium: true,
     location: {
       latitude: 51.35514938496378,
       longitude: 4.673877537499948,
@@ -119,4 +90,4 @@ const offers: Offers = [
   },
 ];
 
-export {offers};
+export {favoriteOffers};
